+++ cps_pp/src/util/dirac_op/d_op_dwf/noarch/dwf_quda.C
@@ -67,7 +67,7 @@ int DiracOpDwf::QudaInvert(Vector *out, Vector *in, Float *true_res, int mat_typ
   QudaGaugeParam gauge_param = newQudaGaugeParam();
   QudaInvertParam inv_param = newQudaInvertParam();
 
-  int f_size_cb = GJP.VolNodeSites() * lat.FsiteSize() / 2;
+  size_t f_size_cb = GJP.VolNodeSites() * lat.FsiteSize() / 2;
 
   //--------------------------------------
   //  Parameter setting for Gauge Data
