+++ cps_pp/src/util/dirac_op/d_op_mobius/d_op_mobius.C
@@ -859,7 +859,7 @@ void DiracOpMobius::CalcHmdForceVecs(Vector *chi)
 
   Vector *chi_new, *rho, *psi, *sigma ;
 
-  int f_size_cb = 12 * GJP.VolNodeSites() * GJP.SnodeSites() ;
+  size_t f_size_cb = 12 * GJP.VolNodeSites() * GJP.SnodeSites() ;
 
   chi_new = f_out ;
