+++ cps_pp/src/util/dirac_op/d_op_wilson_types/d_op_wilson_types.C
@@ -118,7 +118,7 @@ int DiracOpWilsonTypes::RitzLatSize() {
   char *fname = "RitzLatSize()";
   VRB.Func(cname,fname);
 
-  int f_size = GJP.VolNodeSites() * lat.FsiteSize();
+  size_t f_size = GJP.VolNodeSites() * lat.FsiteSize();
 
   switch(dirac_arg->RitzMatOper)
   {
